+++ documentation/docusaurus.config.js
@@ -19,7 +19,6 @@ module.exports = {
         { to: 'docs/introduction/getting-started', label: 'Docs', position: 'left' },
         { to: 'docs/faq', label: 'FAQ', position: 'left' },
         { to: 'docs/contributing', label: 'Contributing', position: 'left' },
-        { to: 'blog', label: 'Blog', position: 'left' },
         {
           href: 'https://github.com/tripss/nestjs-query',
           position: 'right',
@@ -53,10 +52,6 @@ module.exports = {
           title: 'Social',
           items: [
             {
-              label: 'Blog',
-              to: 'blog'
-            },
-            {
               label: 'GitHub',
               href: 'https://github.com/tripss/nestjs-query'
             }
@@ -72,11 +67,6 @@ module.exports = {
         docs: {
           sidebarPath: require.resolve('./sidebars.js'),
           editUrl: 'https://github.com/tripss/nestjs-query/edit/master/documentation/'
-        },
-        blog: {
-          showReadingTime: true,
-          // Please change this to your repo.
-          editUrl: 'https://github.com/tripss/nestjs-query/edit/master/documentation/blog'
         },
         theme: {
           customCss: require.resolve('./src/css/custom.css')